sac cuda can compute models of greater than 100x100 GPU compute elements update multiple cells
Testing on sac-cuda has been limited to models of size 100x100. The limitationis that exactly one cell is computed per GPU compute element. It is important to note that this is limited by the number of available thread blocks. In the revision 47 each GPU compute element must and does perform the computation for multiple cell elements. This is acheived by setting a parameter which specifies the number of xelements and yelements to be computed per gpu processing element.

Tests using the SAC OZT with a 500x500 model and 8x8 elements computed per gpu PE give good agreement to SAC. It is still necessary to compare against the saame model with hyperdiffusion enabled.  It was possible to scale the computation to a 2048x2048 because of the scaling property the model size must be 2^n1 X 2^n2 where n1 and n2 may be different sizes. The maximum number of cells which could be computed per gpu element was found to be 100.