Updated the Rayleigh Taylor problem - further testing required for correct implementation
Allowed the time step to be modified using the maximum wave speed and the Courant condition.
Also need to include contribution from hyperviscosity (important see Caunt and Korpi A&A 369,706-728(2001).

Added 3rd order term to gradops.cu for differential operator this can be sitched on off by defining the compiler switch USE_DORDER3.

Further work required with energy contribution to hyperdiffusion there is also an issue with the boundary terms used for the Orszag-Tang test